backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.6.0
beanie>=1.26.0
pymongo>=4.6.0,<4.12
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
segno==1.6.1
Pillow==10.0.1
email-validator==2.0.0
Jinja2==3.1.2
python-multipart==0.0.6